build: allow to change go build flags
During rpm builds, we want to add the go flags provided by the distro,
but we also need to keep our own flags (in particular the -ldflags which sets
the version in main.go). The problem is that all ldflags must be grouped under
the same -ldflags (later -ldflags overrides the previous ones). This commit
adds GO_LDFLAGS/GO_BUILDFLAGS variables to allow to specify flags in addition
to the ones set by admin-helper Makefile