Command to bind privileged ports
macOS unprivileged programs can't bind 127.0.0.1:(port <= 1024). They
can only bind 0.0.0.0:(port <= 1024).
This is an issue with crc, as we don't want to expose the cluster to the
 world.

For Linux, the issue is different. Binding privileged ports requires an
 extra capabilities.
We have already the admin-helper, just use it to bind privileged ports.

The code is not really nice: we cannot send the fd over stdout. We need
a real socket.
The CLI creates a 3 seconds lifetime socket to share the fd.