daemon: Re-read hosts file for each request
This is quite wasteful, but after hostsfile.Clean() has been called,
the parsed hostfile is no longer valid, and adding more data to it
causes some of its content to disappear, see
https://github.com/code-ready/admin-helper/pull/23#issuecomment-971141415

This issue can only be seen when the admin-helper daemon is used, and
when there's a limit to the number of entries per line, which means this
can only be reproduced on Windows.

By reparsing the hostfile each time we get a request, we workaround this
bug.