hosts.Remove: Fix TestRemoveMultiple*SameLine
Removing multiple entries from the same line in one call to
hosts.Remove() can currently result in an out-of-bound access and a
panic.

The removal is done by iterating over the entries, and calling
removeHostFromLine with the index of the entry to remove from the line.
If the line has for example 2 entries, and we want to remove both, we'll
call removeHostFromLine(0) and then removeHostFromLine(1).

However, after the first removal, the line no longer has an entry with
index 1, which causes the out-of-bound access and the panic when we try
to access it.

To avoid this problem, we can loop over the line from the end, even if
we remove entries, the index of the ones we remove afterwards will still
be valid as they are before the ones we removed.