hosts.Remove: Simplify host deduplication code
For some reason, when deduplicating the hosts passed to hosts.Remove, we
were first generating a map of bool and then generating a map of
struct{}. The 2 maps would have the same set of keys, corresponding
to the list of hosts with no duplicates.

The second map can be generated directly without generating the
uniqueHosts map first.