hosts.Remove: Fix line removal with multiple hosts args
If hosts.Remove() is called with multiple arguments, and the removal
triggers the deletion of an empty line, then the removal of subsequent
host entries can cause out of bounds accesses, as we access the line by
index, and we don't reset our indexes after removing the line.

It's a similar issue to the one fixed in "Fix TestRemoveMultiple*SameLine"
except that it occurs when iterating over the lines instead of iterating
over the hosts on a line.

It's fixed in a similar way, by iterating backwards so that even if we
remove a line, this does not invalidate the indexes of the lines we'll
check after the removed line.