Revert "Fix slice bounds out of range error on remove host entries"
This reverts commit e336a9e26cd76d2c49d23fe62ab958d86bf2a94d.

This keeps the added testcase, but the issue will be fixed differently
in one of the following commits.

The current fix is fine, even if it's slightly wasteful as it iterates N
times over the same line to try to remove hosts entries from it.
The main reason for changing it is that this series will fix a second
bug, and this will allow to keep consistency between the way we iterate
over all the lines, and how we iterate over all the hosts in a given
line.