build: Update golangci-lint to newer version
The version of golangci-lint currently used by admin-helper does not
work with newer go versions, and for some reason we do not get
dependabot updates for golangci-lint.

This commit updates the golangci-lint version used. It's not the latest
as the latest ones require go 1.23, and we need to keep support for 1.22
for a bit longer.